extract Resource model into shared crate
The intention is to have common models for the "frontend" (leptos) and
"backend" (tauri) so that we don't have to repeat this logic.

This involved using cargo workspaces to share dependencies across every
member project by putting shared attributes at the top level.

We also ultimately decided to keep developing commands on the tauri side
as if it were a "backend" server because it can manage the database
connection pool more clearly than leptos can.