pass Vec<Resource> from tauri to leptos
the xml is now properly deserialized on both ends and we can map over
them to produce ui components.

we also noticed that not every result from BGG has a valid year
published, so we had to change the model to Option<i32> (Option<i64>
didn't seem to work) and change how None values were handled at every
step.

also we moved the migrations to shared/models and modified the original
migration to allow NULL in the database for resource.year_published