fix: reactively set tracked on frontend correctly
When Tauri "invoke" generates Javascript objects, Rust "None" is
converted to "null", which is not the type we expected when defining our
models in Javascript. null !== undefined because they are not the same
type, but null == undefined (no type check) because they are both
supposed to be falsy.

This ensures the bookmark symbol is synced to the tracked status (filled
when tracked, empty when not).