fix: no results in search query shouldn't error
The XML returned by the BGG API will not have inner <item> tags when
there are no results found, so we needed to make the "item" attribute of
SearchItems and ThingItems an Option type so that quick_xml knows that a
missing <item> tag means a None value for "item". We provide a default
of an empty Vec (Vec::new) in the case of None with `unwrap_or_else`.

Also, `clone` was not needed :D