feat: combine search_bgg and bgg_list_things
We have a single command now that will call bgg's /thing api with every
chunk of 20 results from /search. The command will always return a tuple
of resources and strings (errors), which will be empty if there are
none, respectively. This avoids having to call the /thing api for each
page of the search results we navigate to.