fix: get all tests passing again
sqlx::test will only run migrations if there is a migrations folder in
the rust project root directory, so we reverted the refactoring split of
database from backend to consolidate these into just the backend
directory.

also got rid of the ugly ../.. references to sqlx fixtures by putting
the fixtures directory next to the source files that refer to it.

also needed up to upgrade sqlx for some reason... RUST!